Various improvements to UI and performance
Streamlined the UI. Opening the app instantly connects to discord.
Can leave, rejoin, and host lobbies without closing the app.

Improved performance. We do not allocate vector per UDP packet.
Mapping IPs to Discord Peers is done with 2 array lookups
instead of std::map search.

We never broadcast and always directly send to the correct peer.
Previously, there was a peer cache that was not used.
Now the peer cache is always kept up to date.